urdu_poem_app/requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
pytesseract==0.3.10